--- component-test/src/main/java/org/apache/fineract/cn/notification/NotificationApiDocumentation.java
@@ -1,289 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.fineract.cn.notification;
-
-import com.google.gson.Gson;
-import org.apache.fineract.cn.notification.api.v1.client.NotificationManager;
-import org.apache.fineract.cn.notification.api.v1.domain.EmailConfiguration;
-import org.apache.fineract.cn.notification.api.v1.domain.SMSConfiguration;
-import org.apache.fineract.cn.notification.api.v1.events.NotificationEventConstants;
-import org.apache.fineract.cn.notification.util.DomainObjectGenerator;
-import org.junit.*;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.restdocs.JUnitRestDocumentation;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.*;
-import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.*;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
-import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SuppressWarnings("Unused")
-public class NotificationApiDocumentation extends TestNotification {
-	
-	@Rule
-	public final JUnitRestDocumentation restDocumentation = new JUnitRestDocumentation("build/doc/generated-snippets/test-notification");
-	
-	private final Gson serialize = new Gson();
-	private MockMvc mockMvc;
-	
-	@Autowired
-	NotificationManager notificationManager;
-	@Autowired
-	private WebApplicationContext context;
-	
-	public NotificationApiDocumentation() {
-		super();
-	}
-	
-	@Before
-	public void setUp() {
-		this.mockMvc = MockMvcBuilders.webAppContextSetup(this.context)
-				.apply(documentationConfiguration(this.restDocumentation))
-				.alwaysDo(document("{method-name}", preprocessRequest(prettyPrint()), preprocessResponse(prettyPrint())))
-				.build();
-	}
-	
-	@Test
-	public void documentCreateSMSConfiguration() throws Exception {
-		final SMSConfiguration randomSMSConfiguration = DomainObjectGenerator.smsConfiguration();
-		
-		this.mockMvc.perform(post("/configuration/sms/create")
-				.accept(MediaType.APPLICATION_JSON_VALUE)
-				.contentType(MediaType.APPLICATION_JSON_VALUE)
-				.content(serialize.toJson(randomSMSConfiguration)))
-				.andExpect(status().isCreated())
-				.andDo(document("document-create-sms-configuration", preprocessRequest(prettyPrint()),
-						requestFields(
-								fieldWithPath("identifier").description("Configuration Id for SMS Gateway"),
-								fieldWithPath("auth_token").description("SMS API authentication token"),
-								fieldWithPath("account_sid").description("SMS API account SID"),
-								fieldWithPath("sender_number").description("SMS API sender number"),
-								fieldWithPath("state").description("The state of the Gateway" +
-										"\n ACTIVE for Gateway to be used" +
-										"\n DEACTIVATED for inactive gateways")
-						)));
-	}
-	
-	@Test
-	public void documentCreateEmailConfiguration() throws Exception {
-		final EmailConfiguration randomEmailConfiguration = DomainObjectGenerator.emailConfiguration();
-		
-		this.mockMvc.perform(post("/configuration/email/create")
-				.accept(MediaType.APPLICATION_JSON_VALUE)
-				.contentType(MediaType.APPLICATION_JSON_VALUE)
-				.content(serialize.toJson(randomEmailConfiguration)))
-				.andExpect(status().isCreated())
-				.andDo(document("document-create-email-configuration", preprocessRequest(prettyPrint()),
-						requestFields(
-								fieldWithPath("identifier").description("Configuration Id for Email Server"),
-								fieldWithPath("host").description("Email Server host address"),
-								fieldWithPath("port").description("Email Server port number"),
-								fieldWithPath("protocol").description("Type of protocol in use " +
-										"\nSMTP" +
-										"\nPOP3" +
-										"\nIMAP"),
-								fieldWithPath("username").description("Email address"),
-								fieldWithPath("app_password").description("Email app password or normal password"),
-								fieldWithPath("smtp_auth").description("Enable SMTP"),
-								fieldWithPath("start_tls").description("Enable TLS"),
-								fieldWithPath("state").description("" +
-										"\n ACTIVE for Gateway to be used" +
-										"\n DEACTIVATED for inactive gateways")
-						)));
-	}
-	
-	@Test
-	public void documentFindSMSConfiguration() throws Exception {
-		final SMSConfiguration randomSMSConfiguration = DomainObjectGenerator.smsConfiguration();
-		
-		this.notificationManager.createSMSConfiguration(randomSMSConfiguration);
-		
-		super.eventRecorder.wait(NotificationEventConstants.POST_SMS_CONFIGURATION, randomSMSConfiguration.getIdentifier());
-		
-		notificationManager.findSMSConfigurationByIdentifier(randomSMSConfiguration.getIdentifier());
-		
-		this.mockMvc.perform(get("/configuration/sms/" + randomSMSConfiguration.getIdentifier())
-				.accept(MediaType.APPLICATION_JSON_VALUE)
-				.contentType(MediaType.APPLICATION_JSON_VALUE)
-				.content(serialize.toJson(randomSMSConfiguration.getIdentifier())))
-				.andExpect(status().isOk())
-				.andDo(document("document-find-sms-configuration", preprocessRequest(prettyPrint()),
-						responseFields(
-								fieldWithPath("identifier").description("Configuration Id for SMS Gateway"),
-								fieldWithPath("auth_token").description("SMS API authentication token"),
-								fieldWithPath("account_sid").description("SMS API account SID"),
-								fieldWithPath("sender_number").description("SMS API sender number"),
-								fieldWithPath("type").description("Type of Configuration"),
-								fieldWithPath("state").description("The state of the Gateway" +
-										"\n ACTIVE for Gateway to be used" +
-										"\n DEACTIVATED for inactive gateways")
-						)));
-	}
-	
-	@Test
-	public void documentFindEmailConfiguration() throws Exception {
-		final EmailConfiguration randomEmailConfiguration = DomainObjectGenerator.emailConfiguration();
-		
-		this.notificationManager.createEmailConfiguration(randomEmailConfiguration);
-		
-		super.eventRecorder.wait(NotificationEventConstants.POST_EMAIL_CONFIGURATION, randomEmailConfiguration.getIdentifier());
-		
-		this.notificationManager.findEmailConfigurationByIdentifier(randomEmailConfiguration.getIdentifier());
-		
-		this.mockMvc.perform(get("/configuration/email/" + randomEmailConfiguration.getIdentifier())
-				.accept(MediaType.APPLICATION_JSON_VALUE)
-				.contentType(MediaType.APPLICATION_JSON_VALUE)
-				.content(serialize.toJson(randomEmailConfiguration.getIdentifier())))
-				.andExpect(status().isOk())
-				.andDo(document("document-find-email-configuration", preprocessRequest(prettyPrint()),
-						responseFields(
-								fieldWithPath("identifier").description("Configuration Id for Email Server"),
-								fieldWithPath("host").description("Email Server host address"),
-								fieldWithPath("port").description("Email Server port number"),
-								fieldWithPath("protocol").description("Type of protocol being used by server " +
-										"\nSMTP" +
-										"\nPOP3" +
-										"\nIMAP"),
-								fieldWithPath("username").description("Email address"),
-								fieldWithPath("app_password").description("Email app password or normal password"),
-								fieldWithPath("smtp_auth").description("Enable SMTP"),
-								fieldWithPath("start_tls").description("Enable TLS"),
-								fieldWithPath("state").description("" +
-										"\n ACTIVE for Gateway to be used" +
-										"\n DEACTIVATED for inactive gateways")
-						)));
-	}
-	
-	@Test
-	public void documentUpdateEmailConfiguration() throws Exception {
-		
-		final EmailConfiguration randomEmailConfiguration = DomainObjectGenerator.emailConfiguration();
-		final EmailConfiguration newRandomConfiguration = DomainObjectGenerator.emailConfiguration();
-		
-		this.notificationManager.createEmailConfiguration(randomEmailConfiguration);
-		
-		super.eventRecorder.wait(NotificationEventConstants.POST_EMAIL_CONFIGURATION, randomEmailConfiguration.getIdentifier());
-		
-		newRandomConfiguration.setIdentifier(randomEmailConfiguration.getIdentifier());
-		newRandomConfiguration.setHost("new.host.com");
-		newRandomConfiguration.setApp_password("changePassword");
-		newRandomConfiguration.setPort("554");
-		newRandomConfiguration.setProtocol("pop3");
-		newRandomConfiguration.setUsername("newaddress@example.com");
-		newRandomConfiguration.setSmtp_auth("false");
-		newRandomConfiguration.setStart_tls("false");
-		newRandomConfiguration.setState("ACTIVE");
-		
-		notificationManager.updateEmailConfiguration(newRandomConfiguration);
-		
-		super.eventRecorder.wait(NotificationEventConstants.UPDATE_EMAIL_CONFIGURATION, newRandomConfiguration.getIdentifier());
-		
-		this.mockMvc.perform(put("/configuration/sms/update")
-				.accept(MediaType.APPLICATION_JSON_VALUE)
-				.contentType(MediaType.APPLICATION_JSON_VALUE)
-				.content(serialize.toJson(newRandomConfiguration)))
-				.andExpect(status().isAccepted())
-				.andDo(document("document-update-email-configuration", preprocessRequest(prettyPrint()),
-						requestFields(
-								fieldWithPath("identifier").description("Configuration Id for Email Server"),
-								fieldWithPath("host").description("Email Server host address"),
-								fieldWithPath("port").description("Email Server port number"),
-								fieldWithPath("protocol").description("Type of protocol in use " +
-										"\nSMTP" +
-										"\nPOP3" +
-										"\nIMAP"),
-								fieldWithPath("username").description("Email address"),
-								fieldWithPath("app_password").description("Email app password or normal password"),
-								fieldWithPath("smtp_auth").description("Enable SMTP"),
-								fieldWithPath("start_tls").description("Enable TLS"),
-								fieldWithPath("state").description("" +
-										"\n ACTIVE for Gateway to be used" +
-										"\n DEACTIVATED for inactive gateways")
-						)));
-	}
-	
-	@Test
-	public void documentUpdateSMSConfiguration() throws Exception {
-		final SMSConfiguration newRandomConfiguration = DomainObjectGenerator.smsConfiguration();
-		final SMSConfiguration randomSMSConfiguration = DomainObjectGenerator.smsConfiguration();
-		
-		this.notificationManager.createSMSConfiguration(randomSMSConfiguration);
-		
-		super.eventRecorder.wait(NotificationEventConstants.POST_SMS_CONFIGURATION, randomSMSConfiguration.getIdentifier());
-		
-		newRandomConfiguration.setIdentifier(randomSMSConfiguration.getIdentifier());
-		newRandomConfiguration.setSender_number("new.host.com");
-		newRandomConfiguration.setState("ACTIVE");
-		newRandomConfiguration.setAccount_sid("asdLAKSKFssdfasdf554");
-		newRandomConfiguration.setAuth_token("aalkeifjlasdfalje333");
-		
-		notificationManager.updateSMSConfiguration(randomSMSConfiguration);
-		
-		this.mockMvc.perform(put("/configuration/sms/update")
-				.accept(MediaType.APPLICATION_JSON_VALUE)
-				.contentType(MediaType.APPLICATION_JSON_VALUE)
-				.content(serialize.toJson(randomSMSConfiguration)))
-				.andExpect(status().isAccepted())
-				.andDo(document("document-update-sms-configuration", preprocessRequest(prettyPrint()),
-						requestFields(
-								fieldWithPath("identifier").description("Configuration Id for SMS Gateway"),
-								fieldWithPath("auth_token").description("SMS API authentication token"),
-								fieldWithPath("account_sid").description("SMS API account SID"),
-								fieldWithPath("sender_number").description("SMS API sender number"),
-								fieldWithPath("state").description("The state of the Gateway" +
-										"\n ACTIVE for Gateway to be used" +
-										"\n DEACTIVATED for inactive gateways")
-						)));
-	}
-	
-	@Test
-	public void documentDeleteSMSConfiguration() throws Exception {
-		final SMSConfiguration randomSMSConfiguration = DomainObjectGenerator.smsConfiguration();
-		
-		this.notificationManager.createSMSConfiguration(randomSMSConfiguration);
-		
-		super.eventRecorder.wait(NotificationEventConstants.POST_SMS_CONFIGURATION, randomSMSConfiguration.getIdentifier());
-		
-		System.out.println(randomSMSConfiguration.getIdentifier());
-		this.mockMvc.perform(delete("/configuration/sms/delete/" + randomSMSConfiguration.getIdentifier())
-				.accept(MediaType.APPLICATION_JSON_VALUE)
-				.contentType(MediaType.APPLICATION_JSON_VALUE))
-				.andExpect(status().isOk())
-				.andDo(document("document-delete-sms-configuration", preprocessRequest(prettyPrint())));
-	}
-	
-	@Test
-	public void documentDeleteEmailConfiguration() throws Exception {
-		final EmailConfiguration randomConfiguration = DomainObjectGenerator.emailConfiguration();
-		
-		notificationManager.createEmailConfiguration(randomConfiguration);
-		super.eventRecorder.wait(NotificationEventConstants.DELETE_EMAIL_CONFIGURATION, randomConfiguration.getIdentifier());
-		
-		this.mockMvc.perform(delete("/configuration/email/delete/" + randomConfiguration.getIdentifier())
-				.accept(MediaType.APPLICATION_JSON_VALUE)
-				.contentType(MediaType.APPLICATION_JSON_VALUE))
-				.andExpect(status().isOk())
-				.andDo(document("document-delete-email-configuration", preprocessResponse(prettyPrint())));
-	}
-}
